Move positionRmsFromCat to util.py
This creates consistency with averageRaDecFromCat,
which is also found in util.py.
