+++ python/lsst/validate/drp/matchreduce.py
@@ -35,10 +35,11 @@ from lsst.afw.table import (SourceCatalog, SchemaMapper, Field,
 from lsst.afw.fits.fitsLib import FitsError
 from lsst.validate.base import BlobBase
 
-from .util import (getCcdKeyName, averageRaDecFromCat, sphDist)
+from .util import (averageRaDecFromCat, getCcdKeyName, positionRmsFromCat,
+                   sphDist)
 
 
-__all__ = ['MatchedMultiVisitDataset', 'positionRmsFromCat']
+__all__ = ['MatchedMultiVisitDataset']
 
 
 class MatchedMultiVisitDataset(BlobBase):
@@ -312,20 +313,3 @@ class MatchedMultiVisitDataset(BlobBase):
         # These attributes are not serialized
         self.goodMatches = goodMatches
         self.safeMatches = safeMatches
-
-
-def positionRmsFromCat(cat):
-    """Calculate the RMS for RA, Dec for a set of observations an object.
-
-    Parameters
-    ----------
-    cat -- collection with a .get method
-         for 'coord_ra', 'coord_dec' that returns radians.
-
-    Returns
-    -------
-    pos_rms -- RMS of positions in milliarcsecond.  Float.
-    """
-    ra_avg, dec_avg = averageRaDecFromCat(cat)
-    ra, dec = cat.get('coord_ra'), cat.get('coord_dec')
-    return positionRms(ra_avg, dec_avg, ra, dec)
